:fire: [refactor] type 코드 분리
