+++ src/components/common/Detail/RoomInfo.tsx
@@ -1,5 +1,5 @@
 import { StDetail } from "../../../libs/styles/StDetail";
-import { DetailDataProps } from "../../../typings/DetailData/detail.type";
+import { DetailDataProps } from "../../../typings/detail.type";
 
 function RoomInfo({ estateDetail }: DetailDataProps) {
   return (
