+++ src/hooks/useEstateList.ts
@@ -26,11 +26,14 @@ interface ApiResponse {
   estateList: EstateListData[];
 }
 export default function useUser() {
-  const { data, isLoading }: { data?: ApiResponse; isLoading: boolean } =
-    useQuery(["estateList"], estateApi.getList, {
+  const { data }: { data?: ApiResponse; isLoading: boolean } = useQuery(
+    ["estateList"],
+    estateApi.getList,
+    {
       onError: (error) => {
         console.error(error);
       },
-    });
+    }
+  );
   return { estateList: data?.estateList };
 }
